Add cmp_versions test with both pre fields empty
Comparing semver strings with empty pre-release labels is potentially
tricky: naively, one might check whether the labels are empty one after
the other, but that would result in returning 1 instead of 0 if the
versions are actually equal, because you'd first check the first version
string's pre-release label, and seeing it empty, you'd return 1.

The build labels test case already happens to catch this type of mistake
but has a description which is not helpful in tracking down what the
problem might be. This new one should make it clearer.